# docs/swagger.yaml
definitions:
...
  dto.ProdutoTamanhoListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ProdutoTamanhoResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.ProdutoTamanhoRequest:
    properties:
      created_by:
        type: integer
      descricao:
        maxLength: 255
        type: string
      empresa_filial_id:
        type: integer
      id:
        type: integer
      sigla:
        maxLength: 10
        type: string
      updated_by:
        type: integer
    required:
    - descricao
    - empresa_filial_id
    type: object
  dto.ProdutoTamanhoResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      descricao:
        type: string
      empresa_filial_id:
        type: integer
      empresa_filial_nome:
        type: string
      id:
        type: integer
      sigla:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
  dto.ProdutoVariacaoListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ProdutoVariacaoResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.ProdutoVariacaoRequest:
    properties:
      cor_id:
        type: integer
      created_by:
        description: AUDITORIA
        type: integer
      empresa_filial_id:
        type: integer
      estoque_atual:
        description: Para o estoque inicial
        type: number
      id:
        description: CAMPOS PRINCIPAIS
        type: integer
      preco_adicional:
        type: number
      produto_id:
        type: integer
      sku:
        maxLength: 50
        type: string
      tamanho_id:
        type: integer
      updated_by:
        type: integer
    required:
    - empresa_filial_id
    - produto_id
    - sku
    type: object
  dto.ProdutoVariacaoResponse:
    properties:
      cor_id:
        type: integer
      cor_nome:
        type: string
      cor_sigla:
        type: string
      created_at:
        description: AUDITORIA
        type: string
      created_by:
        type: integer
      deleted_at:
        type: string
      empresa_filial_id:
        type: integer
      empresa_filial_nome:
        type: string
      estoque_atual:
        type: number
      id:
        description: DADOS PRINCIPAIS
        type: integer
      preco_adicional:
        type: number
      produto_id:
        type: integer
      produto_nome:
        description: RELACIONAMENTOS (apenas IDs e Nomes para evitar recursão)
        type: string
      sku:
        type: string
      tamanho_id:
        type: integer
      tamanho_nome:
        type: string
      tamanho_sigla:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /produto-variacoes:
    get:
      consumes:
      - application/json
      description: Retorna uma lista paginada de variações de produto, com suporte
        a filtros.
      parameters:
      - default: 10
        description: Número de registros por página
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para a paginação
        in: query
        name: offset
        type: integer
      - description: Filtrar por ID do Produto
        in: query
        name: produto_id
        type: integer
      - description: Filtrar por ID da Empresa Filial
        in: query
        name: empresa_filial_id
        type: integer
      - description: Filtrar por ID da Cor
        in: query
        name: cor_id
        type: integer
      - description: Filtrar por ID do Tamanho
        in: query
        name: tamanho_id
        type: integer
      - description: Filtrar por SKU
        in: query
        name: sku
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProdutoVariacaoListResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Lista as variações de produto
      tags:
      - Produto - Variações
    post:
      consumes:
      - application/json
      description: Cria uma nova variação de produto com base nos dados fornecidos.
      parameters:
      - description: Dados para criar a variação de produto
        in: body
        name: variacao
        required: true
        schema:
          $ref: '#/definitions/dto.ProdutoVariacaoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProdutoVariacaoResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Conflito, por exemplo, SKU já existente
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Cria uma nova variação de produto
      tags:
      - Produto - Variações
  /produto-variacoes/{id}:
    delete:
      consumes:
      - application/json
      description: Realiza a exclusão lógica de uma variação de produto.
      parameters:
      - description: ID da Variação de Produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Nenhum conteúdo
        "400":
          description: 'Erro ao excluir (ex: variação com estoque)'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Variação de produto não encontrada
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Exclui uma variação de produto
      tags:
      - Produto - Variações
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de uma variação de produto específica.
      parameters:
      - description: ID da Variação de Produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProdutoVariacaoResponse'
        "400":
          description: ID da variação de produto inválido
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Variação de produto não encontrada
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Busca uma variação de produto por ID
      tags:
      - Produto - Variações
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma variação de produto existente.
      parameters:
      - description: ID da Variação de Produto
        in: path
        name: id
        required: true
        type: integer
      - description: Dados para atualizar a variação de produto
        in: body
        name: variacao
        required: true
        schema:
          $ref: '#/definitions/dto.ProdutoVariacaoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProdutoVariacaoResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Variação de produto não encontrada
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Conflito, por exemplo, SKU já existente
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Atualiza uma variação de produto
      tags:
      - Produto - Variações
...
  /produtos/tamanhos:
    get:
      consumes:
      - application/json
      description: Retorna uma lista paginada de tamanhos de produto, com suporte
        a filtros.
      parameters:
      - description: Número de registros por página
        in: query
        name: limit
        type: integer
      - description: Offset para a paginação
        in: query
        name: offset
        type: integer
      - description: Filtrar por descrição
        in: query
        name: descricao
        type: string
      - description: Filtrar por situação (1=Ativo, 2=Inativo)
        in: query
        name: situacao
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProdutoTamanhoListResponse'
      summary: Lista os tamanhos de produto
      tags:
      - Produto - Tamanhos
    post:
      consumes:
      - application/json
      description: Cria um novo tamanho de produto com base nos dados fornecidos.
      parameters:
      - description: Dados para criar o tamanho de produto
        in: body
        name: tamanho
        required: true
        schema:
          $ref: '#/definitions/dto.ProdutoTamanhoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProdutoTamanhoResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Cria um novo tamanho de produto
      tags:
      - Produto - Tamanhos
  /produtos/tamanhos/{id}:
    delete:
      consumes:
      - application/json
      description: Realiza a exclusão lógica de um tamanho de produto.
      parameters:
      - description: ID do Tamanho de Produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Nenhum conteúdo
        "400":
          description: Erro ao excluir
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Exclui um tamanho de produto
      tags:
      - Produto - Tamanhos
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de um tamanho de produto específico.
      parameters:
      - description: ID do Tamanho de Produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProdutoTamanhoResponse'
        "404":
          description: Tamanho de produto não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Busca um tamanho de produto por ID
      tags:
      - Produto - Tamanhos
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um tamanho de produto existente.
      parameters:
      - description: ID do Tamanho de Produto
        in: path
        name: id
        required: true
        type: integer
      - description: Dados para atualizar o tamanho de produto
        in: body
        name: tamanho
        required: true
        schema:
          $ref: '#/definitions/dto.ProdutoTamanhoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProdutoTamanhoResponse'
        "400":
          description: Erro de validação ou dados inválidos
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Tamanho de produto não encontrado
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Atualiza um tamanho de produto
      tags:
      - Produto - Tamanhos
...
swagger: "2.0"